Footer UI Update Acroos
