+++ src/scenes/about cyberpeace/AboutCyberPeace.jsx
@@ -1,44 +1,9 @@
 import React from 'react';
 import './AboutCyberPeace.css';
 import { motion } from 'framer-motion';
-import { CyberPeaceEarth, CyberPeaceAim, CPFLogo } from '../../assets/photos';
-import {
-  NitiAyog,
-  EndViolence,
-  GCN,
-  ICANN,
-  StalkWare,
-  TaskForce
-} from './partners logos'
+import { CyberPeaceEarth, CyberPeaceAim } from '../../assets/photos';
 
 const AboutCyberPeace = () => {
-
-  const bannerImages = [
-    {
-      title: 'niti-aayog',
-      src: NitiAyog
-    },
-    {
-      title: 'global-compact-network-india',
-      src: GCN
-    },
-    {
-      title: 'icann-at-large',
-      src: ICANN
-    },
-    {
-      title: 'ransomware-task-force',
-      src: TaskForce
-    },
-    {
-      title: 'coalition-against-stalkware',
-      src: StalkWare
-    },
-    {
-      title: 'end-violence-against-children',
-      src: EndViolence
-    }
-  ]
 
   return (
     <div className='cyberpeace'>
@@ -116,30 +81,6 @@ const AboutCyberPeace = () => {
               200 Million netizens across 130 countries.
             </p>
           </motion.div>
-        </div>
-
-        <div className='cyberpeace-partners'>
-          <h2>CyberPeace Community Partners</h2>
-          <div className='banner-container'>
-            <div className='banner'>
-              {bannerImages.map((image, index) => (
-                <img src={image.src} alt={image.title} key={index}/>
-              ))}
-              {bannerImages.map((image, index) => (
-                <img src={image.src} alt={image.title} key={index}/>
-              ))}
-            </div>
-          </div>
-        </div>
-
-
-        <div className='cyberpeace-more'>
-          <div>
-            <img src={CPFLogo} alt="CPF" />
-          </div>
-          <a href="https://www.cyberpeace.org/" target='_blank' rel="noreferrer">
-            Read More
-          </a>
         </div>
     </div>
   )
